Update phoenix and fix resulting compilation errors (#379)
* Update phoenix and fix resulting compilation errors

Due to an issue in how phoenix' CMake script was set up, compilation flags were propagated across targets incorrectly. This was recently fixed and also caused some compilation errors in OpenGothic which this commit addresses.

+++ game/game/aistate.cpp
@@ -14,7 +14,7 @@ AiState::AiState(GameScript& owner,size_t id) {
     mname    = "";
 
     funcIni  = id;
-    funcLoop = -1;
-    funcEnd  = -1;
+    funcLoop = uint32_t(-1);
+    funcEnd  = uint32_t(-1);
     }
   }
